backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=23.0
psycopg2-binary>=2.9
dj-database-url>=2.0
python-dotenv>=1.0
django-cors-headers>=4.0
